vector utilsを追加

+++ src/objs_newnode.c
@@ -6,7 +6,7 @@
 /*   By: tterao <tterao@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/30 16:13:00 by tterao            #+#    #+#             */
-/*   Updated: 2023/10/02 17:01:36 by tterao           ###   ########.fr       */
+/*   Updated: 2023/10/02 17:30:05 by tterao           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ static t_sphere	*_sphere_newnode(const char **info)
 	newnode->coordinate.y = ft_atof(*(info++));
 	newnode->coordinate.z = ft_atof(*(info++));
 	newnode->diameter = ft_atof(*(info++));
-	newnode->radius = newnode->diameter / (float)2.0;
+	newnode->radius = newnode->diameter / 2.0f;
 	newnode->red = ft_atoi(*(info++));
 	newnode->green = ft_atoi(*(info++));
 	newnode->blue = ft_atoi(*(info++));
@@ -62,7 +62,7 @@ static t_cylinder	*_cylinder_newnode(const char **info)
 	newnode->coordinate.y = ft_atof(*(info++));
 	newnode->coordinate.z = ft_atof(*(info++));
 	newnode->diameter = ft_atof(*(info++));
-	newnode->radius = newnode->diameter / (float)2.0;
+	newnode->radius = newnode->diameter / 2.0f;
 	newnode->red = ft_atoi(*(info++));
 	newnode->green = ft_atoi(*(info++));
 	newnode->blue = ft_atoi(*(info++));
